requirements.txt
aiohttp==3.11.16
alembic==1.15.1
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==3.2.0
fastapi==0.115.12
gql== 3.5.2
graphene== 3.4.3
graphene-sqlalchemy==2.1.0
passlib==1.7.4
pika==1.3.2
pydantic==2.10.6
python-multipart==0.0.20
python-jose[cryptography]
python-jose==3.4.0
pyjwt==2.10.1
pytest==8.3.5
redis==5.2.1
requests==2.32.2
sqlalchemy==2.0.39
sqlalchemy[asyncio] 
strawberry-graphql==0.262.5
streamlit==1.41.1
uvicorn==0.34.0